Module Ground.Enum

Enum definitions needed for describing data-encoding ground types.

type assoc = (string * Kaitai.Types.EnumSpec.t) list

An Enum.assoc is an association list of enum id (string) with the corresponding EnumSpec.t.

See Helpers.add_uniq_assoc for handling helper.

val bool_false_name : string

bool_false_name is the string used in the bool enum to represent the value false.

val bool_true_name : string

bool_true_name is the string used in the bool enum to represent the value true.

val bool : string * Kaitai.Types.EnumSpec.t

bool is an association for boolean type.