Module Ground.Type

type definitions needed for describing data-encoding ground types.

type assoc = (string * Kaitai.Types.ClassSpec.t) list

A Type.assoc is an association list of type id (string) with the corresponding ClassSpec.t to be used in a UserType.

See Helpers.add_uniq_assoc for handling helper.

val n_chunk : string * Kaitai.Types.ClassSpec.t

n_chunk is an association for the chunks that compose n.

val n : string * Kaitai.Types.ClassSpec.t

n is an association for n (arbitrarily large natural) type. Requires to also assoc n_chunk.

val z : string * Kaitai.Types.ClassSpec.t

z is an association for z (arbitrarily large integer) type. Requires to also assoc n and n_chunk.

val uint30 : endian:Data_encoding__TzEndian.endianness -> string * Kaitai.Types.ClassSpec.t

uint30 is u4 with a bound check.

val int31 : endian:Data_encoding__TzEndian.endianness -> string * Kaitai.Types.ClassSpec.t

int31 is s4 with a bound check.

val bytes_dyn_uint8 : string * Kaitai.Types.ClassSpec.t

bytes_dyn_uint8 is a byte sequence dynamically sized by a uint8.

val bytes_dyn_uint16 : string * Kaitai.Types.ClassSpec.t

bytes_dyn_uint16 is a byte sequence dynamically sized by a uint16.

val bytes_dyn_uint30 : string * Kaitai.Types.ClassSpec.t

bytes_dyn_uint30 is a byte sequence dynamically sized by a uint30.